Create slope_matrix.py
This program is the updated version of the dT_dH.py.
I kept the earlier version as well to have a reference for the output.
I have the tests to the program which tests if the no. of inputs are correctly provided then the D should always be positive and then the Hliq is positive the lambda is positive. 
A check is added if the off diagonal entries are zero